Rotacion lasers terminada

+++ Anyaroth/Anyaroth/Laser.h
@@ -24,7 +24,7 @@ public:
 
 	virtual void update(const double& deltaTime);
 
-	void Shoot(int angle);
+	void Shoot(double angle);
 	void Stop();
 	virtual void beginCollision(GameObject* other, b2Contact* contact);
 	virtual void endCollision(GameObject* other, b2Contact* contact);
